factored out mini nock lib into its own file
